--- a/CHANGES/1506.misc
+++ b/CHANGES/1506.misc
@@ -1,0 +1,1 @@
+Refactor ``last_sync_task`` in AnsibleRepository model due to performance issues.