fixes: #1506

Issue: [AAH-2367](https://issues.redhat.com/browse/AAH-2367)

Performance testing:

Number of repositories: 493

current performance: 
```
Time
CPU: 14230.32ms (15279.12ms)
SQL: 1663 queries in 1773.93ms
```
![Screenshot from 2023-08-23 13-17-44](https://github.com/pulp/pulp_ansible/assets/19647757/a3684742-b684-4942-9b56-40862a5a2942)

updated solution:
```
Time
CPU: 2972.00ms (3486.00ms)
SQL: 6 queries in 525.13ms
```
![Screenshot from 2023-08-23 13-02-38](https://github.com/pulp/pulp_ansible/assets/19647757/922aeae8-b915-4367-95f3-e57bc9049a1f)