@mdellweg 2023-07-05
Have we considered, not showing that field in the list serializer?
Also do you regularly purge old tasks? The tasking system performance may benefit from it too.