@jerabekjiri 2023-07-07
@mdellweg 
 >Have we considered, not showing that field in the list serializer?
 
 we need it for state, finished_at and error
 
 
![Screenshot from 2023-07-07 11-41-16](https://github.com/pulp/pulp_ansible/assets/19647757/88ee6706-c46f-4e0b-b15c-023de683c158)

>Also do you regularly purge old tasks? The tasking system performance may benefit from it too.
I'm not sure about that I'll ask around