@jerabekjiri 2023-07-10
@mdellweg 
>Also do you regularly purge old tasks? The tasking system performance may benefit from it too.

From Slack conversation: *We may have purged tasks once or twice a long time ago as a one-off thing during an upgrade, but in general no*, so periodically remove stale tasks would most likely help performance