@mdellweg 2023-07-10
I'm unsure what to say about this. First it is duplicating an information already in the db. These tend to get out of sync and be hard to debug. But that is probably not why we should not do this.
The bigger issue is (both with the current situation as well as with this new approach) that the sync task may be deleted/purged at some point. Also what if the last action on the repository was not a sync, but adding/removing content by any other means? I think the best place for this information to live is the info field of the repository version.
https://github.com/pulp/pulpcore/blob/main/pulpcore/app/models/repository.py#L668
If it contains only information about how the repository version was created (without directly referring to the task), it will never get outdated (versions are immutable...).