@jerabekjiri 2023-07-10
@mdellweg thanks for the insight. First, I'd like to test the impact of deleting old tasks in the CRC stage and If this won't be sufficient, adding it to RepositoryVersion sounds like a good idea.