@jerabekjiri 2023-07-21
current solution (@property def last_sync_task):
![Screenshot from 2023-07-21 11-16-30](https://github.com/pulp/pulp_ansible/assets/19647757/358e318e-9c47-4ed3-bff9-ff42938f3b9b)


annotated `last_sync_task` to queryset:
![Screenshot from 2023-07-21 11-23-13](https://github.com/pulp/pulp_ansible/assets/19647757/cd2ec762-282b-4e4f-9c62-6025876897b5)