@jerabekjiri 2023-08-23
Updated title and description to match the code. Adding images showing performance testing:

before:
![Screenshot from 2023-08-23 13-17-44](https://github.com/pulp/pulp_ansible/assets/19647757/a3684742-b684-4942-9b56-40862a5a2942)

after:
![Screenshot from 2023-08-23 13-02-38](https://github.com/pulp/pulp_ansible/assets/19647757/922aeae8-b915-4367-95f3-e57bc9049a1f)