@jerabekjiri 2023-08-24
>Can you explain how the number of queries is still growing when you have 2000 repositories? I would expect the number of queries would still be 6.

you're right, I had the Invalid token error in the environment after rebuilding and didn't notice it affected the queryset. 

created some more repositories:
```
6 repositories:
before: 14 queries in 11.76ms
after: 6 queries in 9.52ms
--------------------------------------------
306 repositories:
before: 214 queries in 248.86ms
after:  6 queries in 107.99ms
--------------------------------------------
606 repositories:
before: 406 queries 473.57ms
after: 6 queries in 47.90ms
--------------------------------------------
906 repositories:
before: 400 queries in 580.72ms
after: 6 queries in 46.41ms
```